requirements.txt
asgiref==3.8.1
astroid==3.2.4
autopep8==2.3.1
black==24.8.0
click==8.1.7
colorama==0.4.6
dill==0.3.8
Django==5.1.1
djangorestframework==3.15.2
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.3
pycodestyle==2.12.1
pylint==3.2.7
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
rest-framework-simplejwt==0.0.2
sqlparse==0.5.1
tomlkit==0.13.2
tzdata==2024.1
